Refactor the ORPC threads code into a submodule.
