+++ ostd/src/orpc/framework/mod.rs
@@ -26,8 +26,9 @@ pub mod errors;
 mod integration_test;
 pub mod notifier;
 pub mod shutdown;
+pub mod threads;
 
-use alloc::{boxed::Box, sync::Weak, vec::Vec};
+use alloc::{sync::Weak, vec::Vec};
 use core::{
     fmt::Display,
     num::NonZeroUsize,
@@ -35,8 +36,7 @@ use core::{
     sync::atomic::{AtomicBool, AtomicUsize, Ordering},
 };
 
-use log::error;
-use spin::Once;
+pub use threads::spawn_thread;
 
 use crate::{
     cpu_local_cell,
@@ -141,64 +141,6 @@ impl ServerBase {
     }
 }
 
-/// The body of a ORPC thread as a closure.
-type ThreadBody = Box<dyn FnOnce() -> Result<(), Box<dyn core::error::Error>> + Send + 'static>;
-
-/// The type of the function used to implement the `spawn_thread` function.
-type SpawnThreadFn = fn(Arc<dyn Server + Send + Sync>, ThreadBody);
-
-/// Injected function for spawning new threads. See [`inject_spawn_thread`].
-static SPAWN_THREAD_FN: Once<SpawnThreadFn> = Once::new();
-
-/// Start a new server thread. This should only be called while spawning a server.
-pub fn spawn_thread<T: Server + Send + 'static>(
-    server: Arc<T>,
-    body: impl (FnOnce() -> Result<(), Box<dyn core::error::Error>>) + Send + 'static,
-) {
-    if let Some(spawn_fn) = SPAWN_THREAD_FN.get() {
-        spawn_fn(server, Box::new(body));
-        return;
-    }
-
-    TaskOptions::new(wrap_server_thread_body(server, Box::new(body)))
-        .spawn()
-        .unwrap();
-}
-
-/// Return a closure wrapping the body of a server thread with the machinery to setup the execution
-/// context and catch and handle errors.
-///
-/// This should only be used by spawn_thread implementations which are injected using
-/// [`inject_spawn_thread`].
-pub fn wrap_server_thread_body(
-    server: Arc<dyn Server + Send + Sync>,
-    body: ThreadBody,
-) -> impl FnOnce() {
-    move || {
-        if let Result::Err(payload) = crate::panic::catch_unwind({
-            let server = server.clone();
-            move || {
-                Server::orpc_server_base(server.as_ref()).attach_task();
-                let _server_context =
-                    CurrentServer::enter_server_context(server.orpc_server_base());
-                if let Result::Err(e) = body() {
-                    Server::orpc_server_base(server.as_ref()).abort(&e);
-                }
-            }
-        }) {
-            let err = errors::RPCError::from_panic(payload);
-            error!("Server thread panicked: {}", err);
-            Server::orpc_server_base(server.as_ref()).abort(&err);
-        }
-    }
-}
-
-/// Set a custom function for spawning threads. This allows overriding the default thread spawning
-/// behavior. This is required in kernels, like Asterinas, that do not run raw OSTD [`Task`]s
-/// correctly.
-pub fn inject_spawn_thread(func: SpawnThreadFn) {
-    SPAWN_THREAD_FN.call_once(|| func);
-}
 /// Methods to access the current server.
 pub struct CurrentServer {
     _private: (),
